requirements.txt
streamlit>=1.31.0
google-genai>=0.2.0
pillow>=10.0.0
python-dotenv>=1.0.0
boto3>=1.26.0
